Move colors to src dir since other components need to use it
